+++ lib/platform/github/index.js
@@ -47,8 +47,6 @@ module.exports = {
   getFile,
   // Commits
   getCommitMessages,
-  // Releases
-  getReleaseList,
 };
 
 // Get all repositories that the user has access to
@@ -1147,23 +1145,6 @@ async function getCommitMessages() {
     return res.body.map(commit => commit.commit.message);
   } catch (err) {
     logger.error({ err }, `getCommitMessages error`);
-    return [];
-  }
-}
-
-async function getReleaseList(repository) {
-  logger.debug('getReleaseList()');
-  try {
-    const res = await get(`repos/${repository}/releases`);
-    return res.body.map(release => ({
-      url: release.html_url,
-      id: release.id,
-      tag: release.tag_name,
-      name: release.name,
-      body: release.body,
-    }));
-  } catch (err) {
-    logger.err({ err }, 'getReleaseList error');
     return [];
   }
 }
