+++ lib/workers/pr/changelog.js
@@ -1,65 +1,11 @@
 const os = require('os');
 const changelog = require('changelog');
 const cacache = require('cacache/en');
+const { addReleaseNotes } = require('./release-notes');
 
 module.exports = {
-  getReleaseNotes,
   getChangeLogJSON,
 };
-
-function massageBody(body) {
-  let massagedBody = body;
-  logger.debug({ massagedBody }, 'pre-massage');
-  // semantic-release cleanup
-  massagedBody = massagedBody.replace(/^<a name=\"[^"]*"><\/a>\n/, '');
-  logger.debug({ massagedBody }, 'after removing a name');
-  massagedBody = massagedBody.replace(
-    /^##? \[[^\]]*\]\(https:\/\/github.com\/[^/]*\/[^/]*\/compare\/.*?\n/,
-    ''
-  );
-  logger.debug({ massagedBody }, 'after removing compare');
-  if (massagedBody !== body) {
-    logger.debug('Cleaned up semantic-release release note');
-  }
-  // np clean-up
-  massagedBody = massagedBody.replace(
-    /https:\/\/github.com\/[^/]+\/[^/]+\/compare\/[^\n]+$/,
-    ''
-  );
-  massagedBody = massagedBody.replace(/\r\n/g, '\n');
-  massagedBody = massagedBody.replace(/\n+$/, '');
-  return massagedBody;
-}
-
-async function getReleaseNotes(repository, version) {
-  logger.debug(`getReleaseNotes(${repository}, ${version})`);
-  const releaseList = await platform.getReleaseList(repository);
-  let releaseNotes;
-  releaseList.forEach(release => {
-    if (release.tag === version || release.tag === `v${version}`) {
-      releaseNotes = release;
-      releaseNotes.body = massageBody(releaseNotes.body);
-      if (!releaseNotes.body.length) {
-        releaseNotes = undefined;
-      }
-    }
-  });
-  logger.debug({ releaseNotes });
-  return releaseNotes;
-}
-
-async function addReleaseNotes(repository, versions) {
-  if (!(repository && versions)) {
-    return versions;
-  }
-  const updatedVersions = [];
-  for (const v of versions) {
-    const releaseNotes = await getReleaseNotes(repository, v.version);
-    logger.debug({ releaseNotes });
-    updatedVersions.push({ ...v, releaseNotes });
-  }
-  return updatedVersions;
-}
 
 async function getChangeLogJSON(depName, fromVersion, newVersion) {
   logger.debug(`getChangeLogJSON(${depName}, ${fromVersion}, ${newVersion})`);
@@ -108,6 +54,9 @@ async function getChangeLogJSON(depName, fromVersion, newVersion) {
       res.project ? res.project.github : null,
       res.versions
     );
+    if (res.versions && res.versions.some(v => v.releaseNotes)) {
+      res.hasReleaseNotes = true;
+    }
     return res;
   } catch (err) {
     logger.debug({ err }, `getChangeLogJSON error`);
